- Redirect to the patient details page after creating a new patient from the medical record page - Display medical history on the patient details page - Allow numeric values for patient vitals instead of only integers - Allow secretary to edit patient vitals even after checking in - Add "Register Patient" button to medical records page - Do not display patient's vitals in patient profile - Fix billing and vitals errors

+++ resources/js/Components/patient-visit-form/PatientSummaryPanel.jsx
@@ -33,7 +33,7 @@ export default function PatientSummaryPanel({
             id: "temperature",
             label: "Temperature",
             value: patient.vitals?.temperature
-                ? parseInt(patient.vitals.temperature) + " °C"
+                ? parseFloat(patient.vitals.temperature) + " °C"
                 : "",
         },
         {
